+++ Changelog.md
@@ -19,6 +19,7 @@
 * Add support of `show_category_name` and `show_series_name` to chart label properties
 * Add method to 3 base formats to check if it contain any user data `#with_data?`
 * Correct parsing of paragraph properties and run properties
+* Paragraphs now have correct parsing of borders
 
 ### Fixes
 * Fix parsing document style id - it can be string, not only digit